rename FastDMRChecker to DMRChecker
